statistics-service/requirements.txt
Flask==2.2.0
flask-cors
psycopg2-binary==2.9.1
werkzeug==2.2.0
python-dotenv
matplotlib
seaborn